Change int to unsigned int

+++ CPart/kmeans.c
@@ -14,8 +14,7 @@ double getDistance(double * point1, double * point2, unsigned int dimNum);
 void copyArrayIntoArray(double ** arrayToChange, double ** arrayToCopy, unsigned int rows, unsigned int cols);
 void freeMatrixMemory(double ** matrixToFree, unsigned int rows);
 double **initialize_centroids(int rows, int cols, int k, FILE * f, double **dataPoints);
-static double **get_new_centroids(int iterations, int rows, int cols, int k, int epsilon, double **dataPoints,
-                            double **centroids);
+static double **get_new_centroids(unsigned int iterations, unsigned int rows, unsigned int cols, unsigned int k, double epsilon, double **dataPoints,double **centroids);
 
 double ** createMatrix(unsigned int rows, unsigned int cols) {
     unsigned int i;
@@ -78,7 +77,7 @@ double ** initialize_centroids(int rows, int cols, int k, FILE *f, double **data
     return centroids;
 }
 
-static double ** get_new_centroids(int iterations, int rows, int cols, int k, int epsilon, double **dataPoints,
+static double ** get_new_centroids(unsigned int iterations, unsigned int rows, unsigned int cols, unsigned int k, double epsilon, double **dataPoints,
                             double **centroids) {
     unsigned int i;
     unsigned int j;
